[#31]refactor: 시간 유효성 검사 로직 재사용 가능하도록 분리
