+++ src/utils/Util.ts
@@ -1,0 +1,27 @@
+import { ERROR } from "../shared/Constants";
+
+const convertTime = (time: string) => {
+  const timeSplit = time.split(":");
+  const hour = Number(timeSplit[0]);
+  const minute = Number(timeSplit[1]);
+  return hour * 60 + minute;
+};
+
+export const isValidTime = (withTime: string, expireTime: string) => {
+  const currentTime = new Date(Date.now() - new Date().getTimezoneOffset() * 60000)
+    .toISOString()
+    .slice(11, 16);
+  const convertCurrentTime = convertTime(currentTime);
+  const convertWithTime = convertTime(withTime);
+  const convertExpireTime = convertTime(expireTime);
+  if (convertCurrentTime >= convertWithTime) {
+    alert(`${ERROR.CREATE.EARLYWITHTIME} ${currentTime}`);
+    return false;
+  } else if (convertWithTime <= convertExpireTime) {
+    alert(`${ERROR.CREATE.LATEEXPIRETIME}`);
+    return false;
+  } else if (convertCurrentTime >= convertExpireTime) {
+    alert(`${ERROR.CREATE.EARLYEXPIRETIME}`);
+    return false;
+  } else return true;
+};
